Parallelise tests with mocha (yarn test:parallel)
This adds a new command to run tests in parallel using mocha.

Mocha executes one thread *per test file* so the runtime is tied to how
many cores you have and how split up the test files are.

On a 32 core machine this cuts the runtime down from 3 minutes to under
1 minute. This could be reduced further by splitting up the test files but
this is beyond the scope of this commit.

More information can be found here:
https://developer.ibm.com/articles/parallel-tests-mocha-v8/